+++ src/cartesiangrid.h
@@ -310,9 +310,9 @@ class CartesianGrid : public BaseCartesianGrid<DEBUG_LEVEL, NDIM>, private Confi
     //    addGhostCells();
     addDiagonalNghbrs();
 
-    for(auto& [name, srf] : m_bndrySurfaces) {
-      srf.updateNeighbors();
-    }
+    //    for(auto& [name, srf] : m_bndrySurfaces) {
+    //      srf.updateNeighbors();
+    //    }
     //    if(m_loadBalancing) {
     //      setWorkload();
     //      calculateOffspringsAndWeights();