+++ src/lbm/moments.h
@@ -44,11 +44,29 @@ template <GInt NDIM, GInt NDIST, LBEquationType EQ, GBool NOSLIP>
 inline void calcDensity_limited(const GInt cellId, const std::set<GInt>& limitedDists, const std::array<GDouble, NDIST>& constants,
                                 const GDouble* normal, const std::function<GDouble&(GInt, GInt)>& fold,
                                 const std::function<GDouble&(GInt, GInt)>& vars) {
+  if(limitedDists.empty()) {
+    // cerr0 << "skipped corner" <<std::endl;
+    return;
+  }
+
   if(EQ == LBEquationType::Navier_Stokes) {
+    const GInt sumC = std::accumulate(constants.begin(), constants.end(), 0);
+
+    const GDouble tmpRho = vars(cellId, LBMVariables<LBEquationType::Navier_Stokes, NDIM>::rho());
     vars(cellId, LBMVariables<LBEquationType::Navier_Stokes, NDIM>::rho()) = 0;
     for(const GInt dist : limitedDists) {
       vars(cellId, LBMVariables<LBEquationType::Navier_Stokes, NDIM>::rho()) += constants[dist] * fold(cellId, dist);
+      //      cerr0 << "dist " << dist << " constants " << constants[dist] << std::endl;
     }
+
+    //    // we are in a corner
+    //        if(sumC < NDIST){
+    //          // assume rho_infty = 1.0
+    //          //todo: this is not necessarily correct probably use a different function for corners
+    //          vars(cellId, LBMVariables<LBEquationType::Navier_Stokes, NDIM>::rho()) += (NDIST-sumC)/static_cast<GDouble>(NDIST) * tmpRho;
+    //          cerr0 << "setting corner fix" << std::endl;
+    //        }
+    //    cerr0 << "rho=" << vars(cellId, LBMVariables<LBEquationType::Navier_Stokes, NDIM>::rho()) << std::endl;
 
     // for no slip walls the velocity in the cell is 0!
     if(!NOSLIP) {